ENTROPY_ENGINE.V3/requirements.txt
jax[cuda12]>=0.4.0
flax>=0.8.0
optax>=0.2.0
brax>=0.10.0
vispy>=0.14.0
pyzmq>=25.0.0
PyQt5>=5.15.0
distrax>=0.1.2
imageio>=2.30.0
imageio-ffmpeg>=0.4.0
hydra-core>=1.3.0
omegaconf>=2.3.0
wandb>=0.16.0
safetensors>=0.4.0
rich>=13.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
